FolderMergerPyQt6/requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
